Use new stopAfterCurrent() API added in amarok 2.3.1.
This allows the script to detect whether to continue playback
after the playlist has finished playing.

Since this removes the necessity to enable or disable the script
manually to work around the lack of integration with the "stop
after current" feature, the enabled check box was moved to the
config dialog, cleaning up amarok's menus a little.

This is still kinda untested, since I'm still running 2.3.0.