Monitor track changes also.
This allows the script to trigger when skipping the last track of
a playlist. The user experience is not the best (amarok will
sometimes start playing the first track again, and to properly detect
when the user actually stops playback, the first milliseconds of the
first track might be played), but it seems to work.

Based on suggestion from kronos003 (irc.freenode.net), although I
didn't use his patch.