Added extra filters
Extra filters added for:
- Not played in the last year
- Playcount = 0
- Played = never
These filters can be enabled in the settings dialog.
Note that they are each an additional filter, but not to each other:
Selecting all three checkboxes will give an album which has at least one
file that meets one of the three additional criteria, not all of them.